Simply API, change rooting strategy.
Roots are now stored in a single global stack of roots, this works because they
are guaranteed to drop in a determined order. The handles created by letroot
just track that the stack needs to be popped.

Heap roots are gone, only stack roots remain, tying every Gc allocation to a
particular stack frame which it is currently guaranteed to live until.

Traits users are not intended to implement directly are moved to the raw
module.